Fix Ledger recipe in response to CQ failures
I had neglected to test all the build variations locally, so some of
the directory names were wrong for arm64.

Also, when I tested this on CQ, the x86-64 release build failed, which I
was unable to reproduce locally, and I noticed that there was no log
output from the QEMU command. I've moved the log line so that it happens
before the failure exception, so that future failures in the QEMU test
step can be diagnosed.

Change-Id: I13915916d28e4f0a25efb23e195ce798618241f8