Use TCP test runner in the Fuchsia recipe
This requires running QEMU in the background, so this change adds
daemonization to the qemu.py wrapper, and a new script that reads the
pidfile and stops the daemon.

The QEMU module and the qemu.py script now also pass through the
necessary flags to do link-local networking with port forwarding.

Since the Fuchsia recipe is now really long, I've split it into several
functions.

To enable local testing, I've added the use_goma property.

Change-Id: I1765a9d2569886d0b82085a2b24fd33304036a38