Run Magenta core tests
Core tests don't start a full system, so they won't work with the
existing technique of using an autorun script with "dm poweroff" at the
end. This change adds a --shutdown_pattern option to qemu.py which
specifies a regex to watch for, which signals that it's safe to shut
down.

This also happens to be a more reliable way to shut down even if
autorun is available, because it forces the script to wait until the
relevant output has actually been flushed. So the existing ("booted")
tests now also shut down this way.

Change-Id: I57c2ae1aae2222d1ad31449247e70cc8bcac7c5d