[sdk] Put the digest of SDK into step name for easier access
The URL of the uploaded archive is published gsutil.upload step, but
since the digest is put into a DEPS file in chromium the full URL link
is less useful. This puts the digest into the step name for the
"upload fuchsia-sdk" step so it shows up on the status page for the
build and can be more easily copy-pasted into the DEPS file.

Change-Id: I5bafcbabb6f07d9439497873f86263b4e4df10e7