Always sleep after run_tests in Fuchsia recipe
Recent test crashes have been missing backtraces because the run_tests
step isn't protected by try-finally.

To avoid nested try-finally blocks, this change also makes it so that
the symbolize step happens before QEMU gets shut down, which is fine
because all it needs is the content of the QEMU log file.

Change-Id: I06fcb622bebddbd4f34c2a84ddb8fb162db07a27