[sdk] Fix jiri.snapshot name to correspond with archive
5866ac658d150906159ca494879d568b0746d1f1 changed the jiri snapshot name
to be a digest of the snapshot contents, which is a meaningless value,
instead of the intended name of the hash of the archive, which is
something consumers of the SDK archive know and care about. This
restores the name of the snapshot so consumers that want to know what
went in to the SDK they are using can identify jiri snapshot
corresponding to the SDK they are using.

Change-Id: Ia9cad944216f56760aae5c718207d5f1df4cf078