Use runtests with autorun for Fuchsia recipe
This is the most straightforward way to get all of the tests running
on the Fuchsia builders. The runtests command simply scans a directory
for test binaries and runs them. Test binaries can be disabled
individually, which causes them to be placed in /system/test/disabled.

This change also simplifies the flow of the recipe by getting rid of the
QEMU background process. People often have trouble responding to build
failures due to the non-linear nature of the Fuchsia recipe. For
example, the "run tests" step produces error messages, but the cause of
those errors is often explained in a later step called "stop qemu and
read log".

When the tests fail, the diagnostic information is going to be in the
middle of a very long QEMU log. Future changes should make use of
gtest's XML output option to put the most important information in the
most visible place.

Some builders (e.g., Ledger, Mozart) have a specified subset of tests
to run. With this change, they will run all tests that are built in
their Fuchsia image. In practice, this means the FTL, MTL, and other
core library tests will start running for their CQ jobs, which shouldn't
be a huge burden.

Change-Id: I81a5e21b89f46d8f0135fef9670733081090ced4