Fix test failure handling
This fixes a bug where the variable step_result in RunTestsWithAutorun
is set to the result of the "symbolize backtraces" step instead of the
"run tests" step when parsing test results. This causes the "run tests"
step to show up in green when tests fail.

To reduce confusion, I've made it so that each step in this function
gets its own variable name. I'm also explicitly storing the failure
reason string for the "run tests" step in a variable, because it can be
set either by looking at the return code or when parsing the QEMU log.

Change-Id: I8ceb54cb861d5a4f764a0c9ae1e92d3fe1e03c3a