Shutdown on a known string after runtests
runtests prints some information about which tests failed after the
"SUMMARY" line, but the recipe is detecting the "SUMMARY" line and
shutting down QEMU before this can happen. So it's hard to tell what
tests failed on the bots.

This change adds an echo line after runtests in the autorun script, and
uses the line as a shutdown trigger, so that the entire output of
runtests can be captured.

Change-Id: I8e99ca92f26983ae7efaacb4648fd678d15b0692