Use the standard bootfs in the Dart recipe
This is a follow-up to https://fuchsia-review.googlesource.com/c/78541/

The previous CL uses the extra_bootdata build argument to add the Dart
test directory to the Fuchsia image. But the recipe was still trying to
build the no-longer-existent generate_dart_test_image target directly
and use its bootfs image.

Change-Id: Icf864c6992e6d43ca7bac868efce4bc98015b97c