Build Zircon host tools with ASan on Fuchsia bots
The new -H switch to //scripts/build-zircon.sh does this.  This
adds several seconds to the build cycle in a way that affects
incremental builds, so it's not on for developers by default.
But that's not a significant slowdown for the overall bot cycle
time, and we want the bots to catch any bugs in these tools.

Change-Id: Ibb73e1f287dd62f211fa13ded91d2d99beeae8ac