fuchsia: Make each test result its own step.
This change creates a new step for each test result and nests each
of these steps inside of a new step called 'tests'. This is the
'correct' way to surface test failures, but right now will result
in many steps. In the future, these steps should be rendered
collapsible in the UI.

Change-Id: I778c8ebcabd4779ebd51ea7fa18cc2f9a3198f43