fuchsia-roller: Remove 'aspirational' code and add docs
This change cleans up 'aspirational' code that was written to show
what the logic would look like if we weren't doing a dry-run with
the new gerrit-client-based code.

As a result, it leaves the recipe in a state where dry_run continues
to not work. FIXMEs and TODOs have been added appropriately, as well
as documentation explaining what the recipe should be doing.

Change-Id: Iae5e9a18005a4f6c85c5b78cdb5e881608056508