fuchsia: Hardcode output block device's PCI address
This change hardcodes the output block device (the MinFS image which
contains test output) to be exposed in QEMU at the PCI address 06.0.
This fixes the TODO to avoid using /dev/class/block/000, which is
inherently non-deterministic in the presence of multiple block devices.

INTK-99 #comment

This change was previously reverted because there was suspicion that it
caused wide-spread infra outages, however there's a high chance this wasn't
the issue since the original issue could be reproduced without these changes.
As a result, this change is being reapplied.

Change-Id: I12035bde7de52e349b5c18344c103bc810cc4180