fuchsia: Use fvm images instead of bootfs
This change updates the fuchsia recipe module to create a qcow2
image from fvm.blk, and to use that resulting image on the test bot
instead of user.bootfs, which is being deprecated.

INTK-99 #comment

This change was previously reverted because there was suspicion that it
caused wide-spread infra outages, however there's a high chance this wasn't
the issue since the original issue could be reproduced without these changes.
As a result, this change is being reapplied.

Change-Id: Ib7d8522e91fc4e0cc8a461c5d7b4902f22fd5d3a