fuchsia: Make topological paths arch-independent
Since ZX-1621 has been finished, the topological paths on x86 and arm64
should now be identical. This change removes _create_runcmds_package'
dependence on the target.

Also, this change exposes that there are no tests which try to execute
tests with arm64, so this change adds a new test case for testing on
arm64.

Change-Id: Ic616945463d2c717b7909b62841719af3b0a8c35