fuchsia: Change symbol directory for test failure
This change fixes the output symbolization directory. In the past, we
were trying to symbolize the test output with the fuchsia build
directory, but that wasn't working because the test output never
contains any backtraces. As a result, we moved to using the kernel
output if any failed tests were detected, but now used zircon's output
directory as the source of debug info. This was also incorrect as
crashes in Fuchsia libraries wouldn't be symbolized correctly. This
change finally corrects this.

Change-Id: I34618825030374a7b507c2db219af9fcbdfd2118