zircon: Enable virtualization in pure QEMU for ARM64
This change adds the "-machine virtualization=true" flag to invocations
of QEMU emulating ARM64 since it starts a guest at EL2, enabling
hypervisor-level tests. We only do this for the QEMU recipe module which
is the only place we run ARM64 QEMU without KVM. We don't support ARM64
QEMU with KVM since nested virtualization with KVM on ARM64 is not currently
available.

INTK-145 #comment

Change-Id: Ieb275cea938e378a8f89ffe7d18ed4662af9d9a7