fuchsia: Downgrade size of MinFS image
This change reduces the size of the MinFS image from 32M to 16M. 16M is
chosen because it is roughly 10x more than we need today, which is a
more reasonable amount of overhead than 32M for today.

Another reason for this change is to force an isolate cache miss, since
evidently what's supposed to be an empty 32M MinFS image actually isn't
empty in all isolate caches (INTK-137). This change effectively acts as
an isolate cache flush across all our bots for the MinFS image.

INTK-137
INTK-184
INTK-188

Change-Id: I4c9db91089afe22df96d3b5bdb31e57492a9d46c