Use hermetic sysroot for linux swift builds
This is less than optimal method for using the sysroot. A lot of
compromises and hacks are involved including trimming down features in
swift that will work with the pass through CMAKE args. Proper support
building with a passed in sysroot is needed to make everything else
work which is still a WIP. Swiftpm, sourcekit, libdispatch,
foundation, and other misc tools have been disabled (nearly anything
that doesn't use CMAKE as it's build system).

Change-Id: Id8667b8b9f5ebd5fb36cf7ca57827f921ddabea2