zircon: Downgrade the MinFS image size
This change downgrades the size of the MinFS image from 32M to 16M
to stay consistent with the sized used in Fuchsia CI. The primary
purpose of this change however is to force a cache miss to avoid issues
with INTK-137, where the current 32M image is "dirty" on some targets.

This should clean up the last of the infra flake caused by INTK-137.

INTK-184

Change-Id: Iff59f6599fafb7ffb9d9f5edc6b8584fa60b6fbc