fuchsia: Update timeout semantics
This change updates the timeout semantics to support recent changes to
the swarming client. More specifically, the swarming client now supports
waiting indefinitely during collect if no timeout is specified. This
change updates the timeout for collect to be indefinite, but explicitly
sets a hard timeout for the swarming task (instead of just using the
default of 1 hour). This way, the swarming client will wait for the task
to reach a terminal state, which guarantees that the swarming client
will always have information about what was happening when the bot timed
out. The next step is to expose this hard timeout as a property in the
recipes that need it.

IN-270

Change-Id: I205c6be01cc56eb7fcace2b57f7bb6078616050f