Use explicit revision when importing Jiri manifest when possible
When triggering a build job, scheduler sends a revision for which the
build was triggered as 'revision' property. However, Jiri was ignoring
that property and always imported the manifest from tip-of-tree. This
means that the revision we're actually building might be potentially
different from the revision we were triggered for.

Jiri was recently extended to support specifying import revision.  This
change modifies all recipes to always pass the explicit revision to the
import command when available. There has also been some cleanup some
unused arguments to ensure consistency across recipes.

Change-Id: I99bd1843841477c3a5410f0b02aa78154d8c2549