[zircon] Refactor isolated test flow
This change refactors the way Zircon tests are invoked in the isolated
case. Note that this does not make any functional change, simply a code
maintainence change.

This change makes a few decisions. It increases some level of code
duplication in the name of some details of the testing process being
more readily available. It also separates the isolated flow into its own
function, to prepare for the device testing flow.

Finally, this change also updates a couple comments to reduce confusion
for future readers.

Change-Id: I2a308b395f4a1ada1b2800ca00036d62fcf27f19