[zircon] Extend sleep time at the beginning of runcmds
This change extends the amount of time we sleep at the beginning of
runcmds to wait for devices to be enumerated by devmgr, since in the
use_isolate case we rely on block devices being available for test
output. The low sleep time previously might have resulted in INTK-237.

This also adds comments pointing to a bug (ZX-1903) which tracks a
feature that could remove the sleep calls.

Change-Id: Ie52e1c196774bc5710870a3c8e5d846e74a28751