[dart_toolchain] Use api.path['cleanup']
The target for api.file.copytree must not exist, so mkdtemp is not the
right thing.  The api.path documentation recommends avoiding ['tmp_base']
and using ['cleanup'] instead, so do that.  The ['cleanup'] semantics do
not require ensuring there's a unique name, since it's always empty when
the recipe runs.

TO-928

Change-Id: I21fabfa880ed0c082f4067c5981dc2656991dcec