[sdk] Upload jiri.snapshot to a location based on the SDK archive hash
This commit restores the extra snapshot-upload behavior that was lost in
213dbce [sdk] Use the fuchsia recipe module

In lieu of any real build metadata indexing, SDK clients rely on this
association to find the jiri.snapshot used to build a specific SDK archive.

As the flip side of this, stop asking api.fuchsia to upload the snapshot to the
location used by other builders.

INTK-251 #done

Change-Id: If2e45198ce3deba17a6d2e7178d8d4a5b32c7323