[qemu] Remove unnecessary old QEMU invocation
This change removes an old QEMU invocation from the QEMU recipe module
in addition to the script which created the invocation. Nothing uses
this any longer, and it likely isn't something we'd want to use in the
future (depends on matching the right regex and killing QEMU forcefully,
provides no path for getting data off of the target).

Change-Id: Ib567eb18b0e66bfb528affa7ded1e284b9a605d3