[zircon] Fix GOMACC when running locally.
Python's lack of type safety strikes again!

The code was written assuming goma_dir is a
recipe_engine.config_types.Path object.
In fact in the local case, it's just a str, and str.join() does something different
than Path.join().
We can't force it to be a Path object in the local case because it's not in a subdir
of one of the recipe's base paths (e.g. [START_DIR]).

Change-Id: If6e593c4d4e2062a929c71b5edfc85fc00867ff2