[gcc_toolchain] Try a different approach to work around static libc++.a
It turns out GCC's configure switch for libraries to use only
affects the build of the compiler itself, not of the build-time
generator programs needed to build it.  Try a different way to
shoehorn in the necessary -lpthread -ldl dependencies that the
static libc++.a fails to get linked in out of the box.

Change-Id: If8ee6816807620f90e6a945a1359cfff55522bec