[api.fuchsia] Use both old and new ways of injecting runcmds
The BLD-19 changes drop support for package details defined entirely
in the package JSON file.  There is now a new GN build argument that
can be used to inject a new package in a different way.

For the BLD-19 changes to land via CQ, the bot recipes need to handle
both old and new methods until the transition is complete.  With the
BLD-19 changes, the old method becomes a no-op so leaving it there
does no harm.

BLD-19 #comment Bots use both old and new ways of injecting runcmds

Change-Id: I3d8a5de8573a533db87365a89dda1334577e049c