[jiri] Use local manifest when running hooks in CQ
Now, hooks will be ran twice, once for the old version and once for the
new version (through update invoked with local manifest after patching)
the change. However, if the old version of hooks is broken, the first
invocation is going to fail and we're never going to get to the second
one. Also, invoking hooks twice is wasteful. So instead invoke hooks
only once with the proper setting of local manifest depending on whether
we're running in CQ or CI.

Change-Id: I13238b16f76d2ae9860ed94c7ee516cffb700229