[clang] Statically link zlib and libxml2
To produce more self-contained toolchain, we build a static version
of zlib and libxml2 and then use them when building Clang.

We avoid the version in our sysroot since those are built as both
shared and static and there isn't an easy way to force CMake to prefer
static version over the shared one. Furthermore, the sysroot copy has
dependency on additional libraries such as liblzma2 and libicu.

TC-137

Change-Id: If01d60fc5ba413e0f91a4b972f88f693f157c70e