[fuchsia_roller] Perform second update from local manifest
This change performs the second jiri update (the one needed to ensure
the source manifest is correct) using the local manifest, rather than
the remote. This ensures that as long as the manifest imported and
updated from are the same, then the source manifest that's created will
include the pin updated by the roller.

IN-249

Change-Id: I847afa2acbde871ee5b11259b7f29d11be1e37e3