[fuchsia] Set Gerrit host for patched snapshot checkout
This change fixes an issue where the Gerrit host used was always the
default in the case of a patched snapshot checkout. Now, it'll use the
Gerrit host it's given.

Also, update tests to more accurately reflect what CQ passes for
patch_gerrit_url, which includes the URL scheme.

IN-420

Change-Id: Ie7f96db15209bf10c38d1cd02d37b881a786e76e