[zircon] Update to new Swarming trigger json output
This change updates the zircon recipe to be compatible with the new
Swarming trigger command's json output format. It also updates the test
API for Swarming so that the Zircon recipe tests pass, which results in
a far-reaching but harmless change (test-only) since everything else
only depends on the collect command understanding the format, and not
the recipe understanding it.

IN-518

Test: Ran the zircon recipe end-to-end locally.

Change-Id: I2eec1a01fb1e6858a67eb814806042b9569c8047