[fuchsia] Use paving workflow by default for devices
This change makes us use the paving workflow by default for all testing
on devices instead of using the netboot.zbi.

For some builders we want to not pave because they may run tests which
are destructive to the underlying block device. For this reason, we need
to make paving optional and only set it at the builder level.

Previously, the change Id7195594f63f1b73aaac98f180eb619dbab7b1b9 was
landed but it caused certain perf testers which destroy the contents of
the paved block device to fail.

Bug: IN-307
Test: Ran locally for fuchsia.py garnet x64 NUC and QEMU configurations.
Change-Id: Ia9cdbc675340e866d32bdedbb809f2eab979c258