[fuchsia][swarming] Add explicit support for EXPIRED
A task may expire if all the bots that would be available to execute it
were alive but busy and could not take on the task.

Currently, we don't handle this explicitly in the fuchsia recipe module,
resulting in hard-to-understand failures unless you dig in. This change
attempts to fill in this whole so that this case is explicitly checked
for.

Change-Id: Iae2a99746bad70ba726a633151d97c37b1973a61