[fuchsia][swarming] Add explicit support for NO_RESOURCE
A task may enter the terminal NO_RESOURCE state if there is obviously no
bot that will be able to run the task. That is, if there are 0 idle and
0 busy bots that are available to run the task. This is distinct from
the EXPIRED state, which indicates that there were some number of busy
bots that COULD have taken up the task, but the task expired while
waiting for one of them to free up.

Change-Id: Ib5fb8fd70444125248115b64a81959d201931308