[recipes] Add new minimal manifest and update test cases.
This will be used by the recipes builder in order to avoid pulling in
any unnecessary code and running unnecessary hooks. Just last week this
builder was broken because of a bug in //infra/prebuilt, but that repo
is not needed for this recipe.

This will also let us optimize the //manifest/infra for developer
friendliness without worrying about breaking that builder.

Bug: IN-245 #comment
Change-Id: I46c6fadfb039a221a896e80a82e4909a893639bd