[fuchsia] Move all build upload logic to api.fuchsia.upload_build_results()
This change consolidates all build-related GCS upload calls to a single callsite.
It prepares the ultimate change to have a single GCS bucket for uploads.

Moreover, it moves the callsite up to right after the build.

One trade-off is the optimization gained from running bloaty in between
spawn_tasks and collect is lost here; however, bloaty runs in only ~1min
and I believe the benefit of this change outweighs the added time.

Bug: IN-691
Change-Id: I116b96e739def7593ab06f768a94b3b4fe51c1d3