[checkout] Patch changes found in $project/.patchfile
This change adds support for a '.patchfile' set in the
root directory of a project.  The feature works as follows:

When the given checkout is derived from a Gerrit patchset (CQ only):
* Checkout the project
* Search for .patchfile
* `jiri patch` every entry in .patchfile.

The patchfile is expected to be a JSON list of objects whose structure
communicates the project, gerrit-host and change ref.

IN-690
IN-768 #comment

Change-Id: Iddef88d6b2695351d681392341fc0bacf195da2d